Checks that RTC is running
Adds a debugging boolean to turn true or false depending on whether it's
necessary to print timer info to the console.
Changed the `loop()` to always `analogWrite()` to `ledPin` and instead
modify the `pinOut` within the if statements. This makes debugging a
little easier, as I can always print `pinOut` to determine the output of
`ledPin`.
Added some arbitrary `Serial.print()`'s that are wrapped in a debugging
if-statement.
Uncommented the `RTC.isrunning()` while-loop in `setup()`. This was
originally commented out because of an issue where the RTC module would
read as though it wasn't running, even though it was telling the time
perfectly fine. After reviewing the DS1307 data sheet, the error
occurred in supplying the RTC module with 3.3V instead of the
recommended 5V. Once this swap was made, no issues have been noticed.