Implements Timestamp class
Creates a Timestamp class which keeps track of the hours and minutes of
the day, as well as the seconds representation of the 24 hour day.
This class is meant to represent the clock time of any day regardless of
the date. It also doesn't keep track of seconds, since precision is not
of utmost importance for the use case of this class.
With this class, the code will be revamped to easily construct
Timestamps for various representations of time, including the alarm
time, the alarm time plus the period, and the future timeout time which
will turn off the LED's at a certain point when there should *ideally*
be enough natural light in the room provided by the irl sun.