Separates Timestamp class into class files
Moved Timestamp class into separate .cpp and .h files.
Gave the class some more functionality and depth by adding setter
methods, adding the non-equivalence (!=) operator, and adding a
`print()` method for easy debugging later on.
Originally setter methods were not added because it was believed that it
would be easier to simply reinstantiate a new Timestamp object, before
needing to reset a Timestamp member variable. This has since been
changed, though the `setDaySeconds()` method is still debatably
necessary, as it holds identical functionality to the Timestamp
constructor.
The non-equivalence (!=) operator should have been added in the previous
commit, but was forgotten. It has since been added.
The `print()` method was originally meant to be created such that a
Timestamp instance could be passed to the `Serial.print()` method, and
print a message containing the Timestamp's current state. However, there
was a lot of unnecessary issues with this approach and ultimately more
time was used on this issue than was needed. The member function
`print()` was a subsequent compromise that still holds the same
functionality as the originally intended implementation.
A final note about Timestamps. The `Timestamp(uint8_t, uint8_t)`
constructor now uses modulus to ensure that the hours and minutes passed
to the constructor, will be confined to the appropriate hours and
minutes on a clock. Originally there was debate over asserting or
throwing an exception if somehow an out-of-bounds integer was passed
into either parameter, but due to this running on an Arduino, I
ultimately decided on a more lenient approach by allowing out-of-bounds
integers to be adjusted to create a functional clock-time timestamp.